[#5993] refacor: Move the JdbcAuthorizationPlugin to authorization-common module
