+++ authorizations/authorization-common/src/main/java/org/apache/gravitino/authorization/common/JdbcMetadataObject.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.gravitino.authorization.jdbc;
+package org.apache.gravitino.authorization.common;
 
 import com.google.common.base.Preconditions;
 import java.util.List;
