+++ authorizations/authorization-common/src/main/java/org/apache/gravitino/authorization/common/JdbcSecurableObject.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.gravitino.authorization.jdbc;
+package org.apache.gravitino.authorization.common;
 
 import java.util.List;
 import org.apache.gravitino.MetadataObject;
